remove duplicates
Entries are sorted by presence and then by name.
If one is online, this is displayed instead of offline.
This solves that one does not need to take all devices with one
to not appear as both online and offline if clicked this way.
fixes https://github.com/HashTec-Potsdam/wer-ist-da/issues/4